backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
imbalanced-learn==0.12.4
torch==2.5.1
joblib==1.4.2
python-multipart==0.0.17
pydantic-settings==2.6.1
pytest==8.3.3
httpx==0.27.2
